rid off ValueParserUtil and move all methods to Maths
