+++ src/main/java/team/creative/creativecore/common/gui/parser/DoubleValueParser.java
@@ -1,14 +1,14 @@
 package team.creative.creativecore.common.gui.parser;
 
 import net.minecraft.network.chat.Component;
-import team.creative.creativecore.common.gui.ValueParserUtil;
+import team.creative.creativecore.common.util.math.Maths;
 
 @FunctionalInterface
 public interface DoubleValueParser {
-    DoubleValueParser NONE = (v, max) -> ValueParserUtil.round(v) + "";
-    DoubleValueParser BLOCKS = (v, max) -> Component.translatable("minecraft.blocks.counting", ValueParserUtil.round(v)).getString();
-    DoubleValueParser PERCENT = (v, max) -> (int) (ValueParserUtil.safeDivide(v, max) * 100.0F) + "%";
-    DoubleValueParser ANGLE = (v, max) -> ValueParserUtil.round(v) + "°";
+    DoubleValueParser NONE = (v, max) -> Maths.safeRound(v) + "";
+    DoubleValueParser BLOCKS = (v, max) -> Component.translatable("minecraft.blocks.counting", Maths.safeRound(v)).getString();
+    DoubleValueParser PERCENT = (v, max) -> (int) (Maths.safeDivide(v, max) * 100.0F) + "%";
+    DoubleValueParser ANGLE = (v, max) -> Maths.safeRound(v) + "°";
 
     String parse(double v, double max);
 }
